a lot of refactoring.
