+++ src/components/D3IndicatorByLine/D3IndicatorByLineChart.js
@@ -3,32 +3,19 @@ import { datastoreContext } from 'layouts/dashboard';
 import React, {
     useContext, useEffect, useRef, useState,
 } from 'react';
-import { byLineMeasureContext, displayDataContext, firstRenderContext } from './ChartContainer';
+import { byLineDisplayDataContext, byLineMeasureContext, displayDataContext, firstRenderContext } from '../D3Container/ChartContainer';
 
 function D3IndicatorByLineChart() {
 
     const D3IndicatorByLineChart = useRef();
 
-    const { datastore, setDatastore } = useContext(datastoreContext);
-    const { displayData, setDisplayData } = useContext(displayDataContext)
-    const { firstRender, setFirstRender } = useContext(firstRenderContext);
-    const { byLineMeasure, setByLineMeasure } = useContext(byLineMeasureContext);
-    const [data, setData] = useState([]);
-    const [memberId, setMemberId] = useState('');
-    const [measurementType, setMeasurementType] = useState('drre');
+    const { byLineDisplayData, setByLinedisplayData } = useContext(byLineDisplayDataContext);
 
     // Binder for react to apply changes to the svg
     const D3IndictaorLineChart = useRef();
 
-    // engage data here
-
     // Date Parser
     const parseDate = d3.timeParse('%Y-%m-%d')
-
-    // Data manipulation
-    const workingList = [];
-    displayData.forEach((item) => workingList.push(item.measure));
-    const measureList = Array.from(new Set(workingList));
 
     // Basic Styling consts to be used later
     const margin = {
@@ -36,7 +23,7 @@ function D3IndicatorByLineChart() {
     };
     const width = (window.innerWidth || document.body.clientWidth) - 100// parseInt(d3.select('#d3-line-chart').style('width'));
     const height = 500;
-    const tickCount = displayData.length / measureList.length;
+    const tickCount = byLineDisplayData.length;
 
     // Clear previous SVG
     d3.select(D3IndictaorLineChart.current).selectAll('*').remove();
@@ -52,7 +39,7 @@ function D3IndicatorByLineChart() {
     // Generates labels and context for x axis
     const x = d3.scaleTime()
         // What data we're measuring
-        .domain(d3.extent(displayData, (d) => parseDate(d.date.split('T')[0])))
+        .domain(d3.extent(byLineDisplayData, (d) => parseDate(d.date.split('T')[0])))
         // The 'width' of the data
         .range([0, width + margin.left]);
 
@@ -62,7 +49,7 @@ function D3IndicatorByLineChart() {
         .call(d3.axisBottom(x).ticks(tickCount).tickFormat(d3.timeFormat('%d-%b-%Y')));
 
     // Generates Label and context for y axis
-    const max = d3.max(displayData, (d) => d.value);
+    const max = d3.max(byLineDisplayData, (d) => d.value);
 
     const y = d3.scaleLinear()
         .domain([0, 5])
@@ -116,28 +103,26 @@ function D3IndicatorByLineChart() {
 
     // Generates the actual line
     const line = d3.line()
-        .curve(d3.curveCardinal)
+        // .curve(d3.curveCardinal)
         .x((d) => x(parseDate(d.date.split('T')[0])))
         .y((d) => y(d.value / 20));
 
     // Iterates through an array variation.
-    if (measureList.length > 0) {
-        measureList.forEach((measure) => {
-            svg.append('path')
-                .datum(displayData.filter((item) => item.measure === measure))
-                .attr('fill', 'none')
-                .attr('stroke', 'black')
-                .attr('opacity', '.33')
-                .attr('stroke-width', 2)
-                .attr('d', line)
-                .on('mouseover', (event) => {
-                    d3.select(event.currentTarget).attr('opacity', '1');
-                })
-                .on('mouseout', (event) => {
-                    d3.select(event.currentTarget).attr('opacity', '.33');
-                });
+
+    svg.append('path')
+        .datum(byLineDisplayData)
+        .attr('fill', 'none')
+        .attr('stroke', 'blue')
+        .attr('opacity', '.33')
+        .attr('stroke-width', 2)
+        .attr('d', line)
+        .on('mouseover', (event) => {
+            d3.select(event.currentTarget).attr('opacity', '1');
+        })
+        .on('mouseout', (event) => {
+            d3.select(event.currentTarget).attr('opacity', '.33');
         });
-    }
+
 
     return (
         <div id="d3-line-chart">
